requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
bleach==6.2.0
blinker==1.4
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
comm==0.2.2
cryptography==3.4.8
datasets==2.19.1
dbus-python==1.2.18
debugpy==1.8.11
decorator==5.1.1
deepspeed==0.14.2
defusedxml==0.7.1
diffusers==0.28.0
dill==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
docstring_parser==0.16
evaluate==0.4.2
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.3.1
grpcio==1.69.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3


matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.11.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1

PyJWT==2.3.0
pynvml==12.0.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
SecretStorage==3.3.1
Send2Trash==1.8.3
sentence-transformers==2.7.0
sentencepiece==0.2.0
shtab==1.7.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.19.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.42.3
triton==2.3.0
trl==0.9.4
typeguard==4.4.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tyro==0.9.8
tzdata==2024.2
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.3.0
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.18.3
zipp==1.0.0
albumentations
opencv-python
mlflow
streamlit
performer-pytorch
isort
blue
timm
toml
gdown
ruamel.yaml==0.18.14